fixed clean-up script

+++ provisioning-scripts/install-parallel.sh
@@ -1,24 +1,30 @@
 #!/usr/bin/env bash
 
-# (wget -O - pi.dk/3 || curl pi.dk/3/) | bash
-# rm -fr .gnupg parallel-*
+CENTOS_VERSION=$(rpm -E %{rhel})
 
-APPLICATION=parallel
-# Last viable version
-VERSION=20190222
-LATEST=${APPLICATION}-${VERSION}
-ARCHIVE=${LATEST}.tar.gz
+if [[ $CENTOS_VERSION -gt 6 ]]; then
+    # (wget -O - pi.dk/3 || curl pi.dk/3/) | bash
+    # rm -fr .gnupg parallel-*
 
-wget http://git.savannah.gnu.org/cgit/parallel.git/snapshot/$ARCHIVE
-tar xzf	$ARCHIVE
-CURRENT_DIR=$(pwd)
-cd $LATEST || exit 2
-./configure && make && make install
+    APPLICATION=parallel
+    # Last viable version
+    VERSION=20190222
+    LATEST=${APPLICATION}-${VERSION}
+    ARCHIVE=${LATEST}.tar.gz
 
-for user_dir in $HOME_DIR /root; do
-    mkdir -p $user_dir/.parallel
-    touch $user_dir/.parallel/will-cite
-done
+    wget http://git.savannah.gnu.org/cgit/parallel.git/snapshot/$ARCHIVE
+    tar xzf	$ARCHIVE
+    CURRENT_DIR=$(pwd)
+    cd $LATEST || exit 2
+    ./configure && make && make install
 
-cd $CURRENT_DIR
-rm -fr $ARCHIVE $LATEST
+    for user_dir in $HOME_DIR /root; do
+        mkdir -p $user_dir/.parallel
+        touch $user_dir/.parallel/will-cite
+    done
+
+    cd $CURRENT_DIR
+    rm -fr $ARCHIVE $LATEST
+else
+    echo "\n[WARN] parallel requires Centos 7 or higher.\n"
+fi
